+++ source/ni/media/alac/codec/ag_enc.c
@@ -2,25 +2,25 @@
  * Copyright (c) 2011 Apple Inc. All rights reserved.
  *
  * @APPLE_APACHE_LICENSE_HEADER_START@
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *     http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- * 
+ *
  * @APPLE_APACHE_LICENSE_HEADER_END@
  */
 
 /*
 	File:		ag_enc.c
-	
+
 	Contains:   Adaptive Golomb encode routines.
 
 	Copyright:	(c) 2001-2011 Apple, Inc.
@@ -94,15 +94,15 @@ static int32_t ALWAYS_INLINE abs_func( int32_t a )
 	int32_t isneg  = a >> 31;
 	int32_t xorval = a ^ isneg;
 	int32_t result = xorval-isneg;
-	
-	return result;	
+
+	return result;
 }
 
 static uint32_t ALWAYS_INLINE read32bit( uint8_t * buffer )
 {
 	// embedded CPUs typically can't read unaligned 32-bit words so just read the bytes
 	uint32_t		value;
-	
+
 	value = ((uint32_t)buffer[0] << 24) | ((uint32_t)buffer[1] << 16) |
 			 ((uint32_t)buffer[2] << 8) | (uint32_t)buffer[3];
 	return value;
@@ -138,10 +138,10 @@ static int32_t dyn_code(int32_t m, int32_t k, int32_t n, uint32_t *outNumBits)
 		if (numBits > MAX_PREFIX_16 + MAX_DATATYPE_BITS_16)
 		{
 		    numBits = MAX_PREFIX_16 + MAX_DATATYPE_BITS_16;
-		    value = (((1<<MAX_PREFIX_16)-1)<<MAX_DATATYPE_BITS_16) + n;            
+		    value = (((1<<MAX_PREFIX_16)-1)<<MAX_DATATYPE_BITS_16) + n;
 		}
 	}
-	
+
 	*outNumBits = numBits;
 
 	return (int32_t) value;
@@ -163,7 +163,7 @@ static int32_t dyn_code_32bit(int32_t maxbits, uint32_t m, uint32_t k, uint32_t 
 
 		de = (mod == 0);
 		numBits = div + k + 1 - de;
-		value = (((1<<div)-1)<<(numBits-div)) + mod + 1 - de;		
+		value = (((1<<div)-1)<<(numBits-div)) + mod + 1 - de;
 		if (numBits > 25)
 			goto codeasescape;
 	}
@@ -176,7 +176,7 @@ codeasescape:
 		*overflowbits = maxbits;
 		didOverflow = 1;
 	}
-	
+
 	*outNumBits = numBits;
 	*outValue = value;
 
@@ -203,7 +203,7 @@ static void ALWAYS_INLINE dyn_jam_noDeref(unsigned char *out, uint32_t bitPos, u
 
 	value  = (value << shift) & mask;
 	value |= curr & ~mask;
-	
+
 	*i = Swap32BtoN( value );
 }
 
@@ -215,7 +215,7 @@ static void ALWAYS_INLINE dyn_jam_noDeref_large(unsigned char *out, uint32_t bit
 	uint32_t	curr;
 	uint32_t	mask;
 	int32_t			shiftvalue = (32 - (bitPos&7) - numBits);
-	
+
 	//Assert(numBits <= 32);
 
 	curr = *i;
@@ -242,7 +242,7 @@ static void ALWAYS_INLINE dyn_jam_noDeref_large(unsigned char *out, uint32_t bit
 		w  = (value << shiftvalue) & mask;
 		w |= curr & ~mask;
 	}
-	
+
 	*i = Swap32BtoN( w );
 }
 
@@ -281,7 +281,7 @@ int32_t dyn_comp( AGParamRecPtr params, int32_t * pc, BitBuffer * bitstream, int
     c=0;
 	status = ALAC_noErr;
 
-    while (c < numSamples)
+    while (c < (uint32_t)numSamples)
     {
         m  = mb >> QBSHIFT;
         k = lg3a(m);
@@ -300,8 +300,8 @@ int32_t dyn_comp( AGParamRecPtr params, int32_t * pc, BitBuffer * bitstream, int
 		if ( dyn_code_32bit(bitSize, m, k, n, &numBits, &value, &overflow, &overflowbits) )
 		{
 			dyn_jam_noDeref(out, bitPos, numBits, value);
-			bitPos += numBits;			
-			dyn_jam_noDeref_large(out, bitPos, overflowbits, overflow);			
+			bitPos += numBits;
+			dyn_jam_noDeref_large(out, bitPos, overflowbits, overflow);
 			bitPos += overflowbits;
 		}
 		else
@@ -309,7 +309,7 @@ int32_t dyn_comp( AGParamRecPtr params, int32_t * pc, BitBuffer * bitstream, int
 			dyn_jam_noDeref(out, bitPos, numBits, value);
 			bitPos += numBits;
 		}
-      
+
         c++;
         if ( rowPos >= rowSize)
         {
@@ -325,14 +325,14 @@ int32_t dyn_comp( AGParamRecPtr params, int32_t * pc, BitBuffer * bitstream, int
 
         zmode = 0;
 
-        RequireAction(c <= numSamples, status = kALAC_ParamError; goto Exit; );
+        RequireAction(c <= (uint32_t)numSamples, status = kALAC_ParamError; goto Exit; );
 
-        if (((mb << MMULSHIFT) < QB) && (c < numSamples))
+        if (((mb << (uint32_t)MMULSHIFT) < (uint32_t)QB) && (c < (uint32_t)numSamples))
         {
             zmode = 1;
             nz = 0;
 
-            while(c<numSamples && *inPtr == 0)
+            while(c<(uint32_t)numSamples && *inPtr == 0)
             {
             	/* Take care of wrap-around globals. */
                 ++inPtr;
